+++ changelog.md
@@ -5,6 +5,7 @@
 - added config options to re-log many exceptions encountered during EMF loading, even some not normally logged due to config settings
 - fixed `spectral_arrow` model breaking in 1.21.3
 - limited `time`, `frame_counter`, `id`, and a few other variables to 27720
+- fixed single block chest models not loading properly in 1.21.2+
 
 [2.3.0]
 - the remaining 1.21.3 changes should be fairly "stable" now and are unlikely to have significant changes going forwards, likely only semantic ones as OptiFine develops